some_warnings_fixed

+++ Container.h
@@ -15,7 +15,7 @@ public:
 
     void ClickLeftEvent      (const Event &event) override
     {
-        for (int i = 0; i < children_.size(); i++)
+        for (size_t i = 0; i < children_.size(); i++)
         {
             children_[i]->ClickLeftEvent(event);
         }
