+++ HSVwindow.h
@@ -99,7 +99,7 @@ public:
         get_start_field().print_value();
         printf("$$$$$$\n");
         
-        for (int i = 0; i < children_.size(); i++)
+        for (size_t i = 0; i < children_.size(); i++)
         {
             children_[i]->ClickLeftEvent(event);
         }
